breakdown of test class on header and cpp
