+++ projects/rocprofiler-compute/src/lib/rocprofiler_compute_tool/test/main.cpp
@@ -1,0 +1,7 @@
+#include "test_rocprifiler_compute_tool.h"
+
+int main(int argc, char** argv)
+{
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
